RFC2865 states that Nas-Port attribute is 32bits unsigned integer. 	PostgreSQL doesn't know unsigned integers. 	The PostgreSQL INTEGER type is 32bits signed. 	The PostgreSQL BIGINT type is 64bits signed. 	Changing the NASPortId field to BIGINT accomodates for 32bits 	unsigned values from the Nas-Port attributes. (closes #188)
	Thanks to Thor Spruyt for the patch.

+++ src/modules/rlm_sql/drivers/rlm_sql_postgresql/db_postgresql.sql
@@ -20,7 +20,7 @@ CREATE TABLE radacct (
 	UserName		VARCHAR(64),
 	Realm			VARCHAR(64),
 	NASIPAddress		INET NOT NULL,
-	NASPortId		INTEGER,
+	NASPortId		BIGINT,
 	NASPortType		VARCHAR(32),
 	AcctStartTime		TIMESTAMP with time zone,
 	AcctStopTime		TIMESTAMP with time zone,
